Correcoes na pontuacao.

+++ software/serial.c.h
@@ -174,7 +174,7 @@ void Serial_Log (void) {
     Serial.println();
 
     u16 avg = (G.jogs[0].total + G.jogs[1].total) / 2;
-    u16 equ = (!S.equilibrio ? 0 : avg - min(avg, min(G.jogs[0].total,G.jogs[1].total)*11/10));
+    u16 equ = (!S.equilibrio ? 0 : avg - min(avg, ((u32)min(G.jogs[0].total,G.jogs[1].total))*11/10));
 
     sprintf_P(STR, PSTR("Media ........... %02d.%02d"), avg/100, avg%100);
     Serial.print(STR);
